fix: 플레이리스트 기본 커버 s3 삭제 오류 수정

+++ backend/daumnal/src/main/java/com/ssafy/daumnal/music/service/impl/PlaylistServiceImpl.java
@@ -253,6 +253,8 @@ public class PlaylistServiceImpl implements PlaylistService {
             throw new NotSameException(NOT_SAME_LOGIN_MEMBER_AND_PLAYLIST_OWNER);
         }
         playlistRepository.deleteById(playlistId);
-        s3Service.delete(playlist.getCoverUrl());
+        if (!playlist.getCoverUrl().equals(PLAYLIST_DEFAULT_COVER_URL)) {
+            s3Service.delete(playlist.getCoverUrl());
+        }
     }
 }
